+++ packages/expo-modules-core/expo-module-gradle-plugin/src/main/kotlin/expo/modules/plugin/ProjectConfiguration.kt
@@ -27,7 +27,10 @@ internal fun Project.applyKotlin(kotlinVersion: String, kspVersion: String) {
 }
 
 internal fun Project.applyDefaultDependencies() {
-  project.dependencies.add("implementation", project.project(":expo-modules-core"))
+  val modulesCore = project.project(":expo-modules-core")
+  if (project != modulesCore) {
+    project.dependencies.add("implementation", project.project(":expo-modules-core"))
+  }
 }
 
 internal fun Project.applyDefaultAndroidSdkVersions() {